pyproject.toml
[project]
name = "jobs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=2.2.1",
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.18.0",
    "openai>=2.6.0",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
